fix(core): address integer edge cases and precision issues
- Avoid Int overflow in `GridSpec` by using Long for bounds computation
- Handle `Int.MIN_VALUE` explicitly in NumberComposer since `absoluteValue` overflows and returns a negative value
- Use `roundToInt()` for corner cache keys to avoid precision loss from truncation
- Fix `CornerStyle.lerp` to switch shape at `t=0.5` instead of always using `end.shape`
- Correct `Random.nextInt` usage to cover full `0..9999` range

Refs: #8, #10

+++ numberbricks/src/main/kotlin/io/github/beankitk/numberbricks/core/geometry/GridSpec.kt
@@ -38,7 +38,7 @@ fun GridSpec(rows: Int, cols: Int, brickCount: Int): GridSpec {
     require(rows > 0) { "Rows must be greater than 0" }
     require(cols > 0) { "Cols must be greater than 0" }
     require(brickCount > 0) { "Brick count must be greater than 0" }
-    require(brickCount <= rows * cols) { "Brick count cannot be greater than ${rows * cols}" }
+    require(brickCount.toLong() <= rows.toLong() * cols.toLong()) { "Brick count cannot be greater than ${rows * cols}" }
 
     return object : GridSpec {
         override val rows = rows
